Release port for non listening sockets which are bound with `any` address.
When a TCP client socket uses the wildcard("any") address for binding, the port
is reserved across both the IPv4 and IPv6 protocol families. During close() or
abort() calls we were only releasing the port reservation corresponding to the
established IPv6 connection tuple, failing to release the IPv4 reservation and
thus blocking future IPv4 bindings to that port.

The fix implements a mechanism to track if a socket was bound to the "any"
address. This is used during socket close/abort to release all relevant port
reservations. It also corrects the behavior to preserve the socket's network
protocol settings, when a bind had been performed on the client socket before
connect.A syscall test has been added to verify the behavior during close.

PiperOrigin-RevId: 749282812